KADANE'S PROBLRM FIX FOR ALL NEGATIVE NUMBERS ARRAY
Problem:
If all elements in the array are negative (e.g., {-5, -3, -1}),  code would output 0, which is incorrect. The correct maximum subarray sum in such a case is the largest (least negative) number, which is -1.

Fix:
Track the maximum sum without forcing currsum to zero in the beginning and only reset currsum when necessary after updating maxsum.